third_party/blink/web_tests/platform/generic/fast/dom/DOMImplementation/createDocument-XMLDocument-expected.txt

Check that DOMImplementation.createDocument() returns an XMLDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.XMLDocument.name is "XMLDocument"
PASS XMLDocument.prototype.__proto__ is Document.prototype
PASS document.implementation.createDocument("", "").__proto__ is XMLDocument.prototype
PASS document.implementation.createDocument("", "html").__proto__ is XMLDocument.prototype
PASS document.implementation.createDocument("http://www.w3.org/1999/xhtml", "html").__proto__ is XMLDocument.prototype
PASS document.implementation.createDocument("http://www.w3.org/2000/xmlns/", "xmlns:test").__proto__ is XMLDocument.prototype
PASS document.implementation.createDocument("http://www.w3.org/2000/svg", "svg").__proto__ is XMLDocument.prototype
PASS successfullyParsed is true

TEST COMPLETE
